+++ Tutorial SFML/Room.cpp
@@ -1,6 +1,6 @@
 #include "Room.h"
 
-Room::Room(const std::string& id, int maxPlayers) : _id(id), _maxPlayers(maxPlayers) {}
+Room::Room(const std::string& id, short maxPlayers) : _id(id), _maxPlayers(maxPlayers) {}
 
 bool Room::IsFull() const { return _players.size() >= _maxPlayers; }
 
@@ -14,4 +14,4 @@ void Room::AddPlayer(Client* client)
 const std::vector<Client*>& Room::GetPlayers() const { return _players; }
 
 std::string Room::GetID() const { return _id; }
-int Room::GetMaxPlayers() const { return _maxPlayers; }
+short Room::GetMaxPlayers() const { return _maxPlayers; }
